cleanup
Removed MP3_DISABLE flag, as the problem is not with actually playing
the MP3s. Moved the actual playing of the MP3s out of the FSM state
handlers, under the idea that it would correct a memory leak, to no
avail. However, I think I like it better this way anyway.